pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chats.py"
version = "0.1.0"
description = "A Python wrapper for writing bots for originchats.mistium.com"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "requests",
    "websockets",
]
license = { file = "LICENSE" }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
